+++ kernel/ide.c
@@ -8,75 +8,19 @@ Disk buffer cache/LBA read/write support
 #include <x86.h>
 #include <mutex.h>
 #include <traps.h>
+#include <ide.h>
 
-#define BLOCK_SIZE		512
-#define SECTOR_SIZE		512
-
-#define B_BUSY	0x1		// buffer is locked by a process
-#define B_VALID	0x2		// buffer has been read from disk
-#define B_DIRTY	0x4		// buffer has been written to
-
-
-/* Define IDE status bits */
-#define IDE_BSY 		(1<<7)	// Drive is preparing to send/receive data
-#define IDE_RDY 		(1<<6)	// Clear when drive is spun down, or after error
-#define IDE_DF			(1<<5)	// Drive Fault error
-#define IDE_ERR			(1<<0)	// Error has occured
-
-#define IDE_IO			0x1F0	// Main IO port
-#define IDE_DATA		0x0 	// R/W PIO data bytes
-#define IDE_FEAT		0x1 	// ATAPI devices
-#define IDE_SECN		0x2 	// # of sectors to R/W
-#define IDE_LOW			0x3 	// CHS/LBA28/LBA48 specific
-#define IDE_MID 		0x4
-#define IDE_HIGH		0x5
-#define IDE_HEAD		0x6 	// Select drive/heaad
-#define IDE_CMD 		0x7 	// Command/status port
-#define IDE_ALT			0x3F6	// alternate status
-#define LBA_LOW(c)		((uint8_t) (c & 0xFF))
-#define LBA_MID(c)		((uint8_t) (c >> 8) & 0xFF)
-#define LBA_HIGH(c)		((uint8_t) (c >> 16) & 0xFF)
-#define LBA_LAST(c)		((uint8_t) (c >> 24) & 0xF)
-
-#define IDE_CMD_READ  0x20
-#define IDE_CMD_WRITE 0x30
-
-
-#define MAX_OP_BLOCKS	16	// Max # of blocks any operation can write (8 KB)
-
-
-void sleep(void* channel, mutex* m) {
-	release(m);
-}
+void sleep(void* channel, mutex* m) {}
 void wake(void* channel) {} 
 
-mutex idelock = {.lock = 0};
-
-typedef struct ide_buffer {
-	uint8_t flags;				// buffer flags
-	uint32_t dev;				// device number
-	uint32_t block;				// block number
-	struct ide_buffer* next;	// next block in queue
-	uint8_t data[BLOCK_SIZE];	// 1 disk sector of data
-} buffer;
-
-// This is a pointer to the buffer being currently processed
-buffer* idequeue;
-
-static int present = 0;
-
-void buffer_init() {
-	idequeue = malloc(sizeof(buffer) * MAX_OP_BLOCKS);	
-	for (buffer* q = idequeue; q < idequeue + MAX_OP_BLOCKS; q++) {
-		memset(q, 0, sizeof(buffer));
-		q->next = idequeue;
-	}
-}
-
-
+mutex idelock = {.lock = 0};		// spinlock for disk access
+buffer* idequeue;					// cache of buffer access
+static int present = 0;				// Dev 1 present?
 
 /*
 Wait for IDE device to become ready
+check =  0, do not check for errors
+check != 0, return -1 if error bit set
 */
 static int ide_wait(int check) {
 	char r;
@@ -105,7 +49,7 @@ static void ide_op(buffer* b) {
 	int sector_per_block = BLOCK_SIZE / SECTOR_SIZE;	// 1
 	int sector = b->block * sector_per_block;
 	ide_wait(0);
-	outb(0x3F6, 0);	// Generate interrupt
+	outb(IDE_ALT, 0);	// Generate interrupt
 	outb(IDE_IO | IDE_SECN, sector_per_block);	// # of sectors
 	outb(IDE_IO | IDE_LOW, LBA_LOW(sector));
 	outb(IDE_IO | IDE_MID, LBA_MID(sector));
@@ -203,11 +147,20 @@ void ide_test() {
 	b->flags = 0;
 	b->dev = 1;
 	b->block = 7;
+
+	buffer* c = malloc(sizeof(buffer));
+	c->flags = 0;
+	c->dev = 1;
+	c->block = 1;
 	printf("RW status: %d\n", ide_rw(b));
+	printf("RW status: %d\n", ide_rw(c));
 
 	int ticks = get_ticks();
-	while(ticks + 10 > get_ticks());
+	while(ticks + 20 > get_ticks());
 
 	for (int i = 0; i < BLOCK_SIZE; i++)
 		printf("%x", b->data[i]);
+	printf("\nBlock2\n");
+	for (int i = 0; i < BLOCK_SIZE; i++)
+		printf("%x", c->data[i]);
 }
